Add module to answer question 1.
The module has the class HumanResources who establishes the conection with the database and manage human resources data. The method create_random_db populates the hhrr database. The method get_avg_children_per_marriage rethrn the average number of children per marriage. The method get_person_with_max_number_grandchildren return a dictionary with the data of the person with the maximum number of grandchildren.